Move imageloader to http package
