+++ app/src/main/java/me/jessyan/mvparms/demo/mvp/ui/holder/UserItemHolder.java
@@ -7,8 +7,8 @@ import android.widget.TextView;
 import com.jess.arms.base.App;
 import com.jess.arms.base.BaseHolder;
 import com.jess.arms.di.component.AppComponent;
-import com.jess.arms.widget.imageloader.ImageLoader;
-import com.jess.arms.widget.imageloader.glide.ImageConfigImpl;
+import com.jess.arms.http.imageloader.ImageLoader;
+import com.jess.arms.http.imageloader.glide.ImageConfigImpl;
 
 import butterknife.BindView;
 import io.reactivex.Observable;
