+++ arms/src/main/java/com/jess/arms/http/imageloader/ImageLoader.java
@@ -1,4 +1,4 @@
-package com.jess.arms.widget.imageloader;
+package com.jess.arms.http.imageloader;
 
 import android.content.Context;
